requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
pillow==10.0.1
pyautogui==0.9.54
numpy==1.24.3
python-string-similarity==0.2.0
pynput==1.7.6
keyboard==0.13.5
mouse==0.7.1
tkinter-tooltip==2.0.0
configparser==5.3.0 